re-add module js libraries
