+++ rpc-tests/tests/0001-transfer-tokens.js
@@ -1,37 +1,8 @@
 async function run(nodeName, networkInfo, args) {
+    const { sendTransaction } = await import("./tx-utils.mjs");
+
     const {wsUri, userDefinedTypes} = networkInfo.nodesByName[nodeName];
     const api = await zombie.connect(wsUri, userDefinedTypes);
-
-    const SPAWNING_TIME = 500000;
-
-    async function sendTransaction(transaction, sender) {
-      return new Promise((resolve, reject) => {
-        let unsubscribe;
-        let timeout;
-
-        transaction.signAndSend(sender, async (result) => {
-          console.log(`Current status is ${result?.status}`);
-
-          if (result.isFinalized) {
-            if (unsubscribe) {
-              unsubscribe();
-            }
-
-            clearTimeout(timeout);
-            resolve(true);
-          }
-        }).then(unsub => {
-          unsubscribe = unsub;
-        }).catch(error => {
-          console.error(error);
-          reject(error);
-        });
-
-        timeout = setTimeout(() => {
-          reject(new Error('Transaction timeout'));
-        }, SPAWNING_TIME);
-      });
-    }
 
     await zombie.util.cryptoWaitReady();
 
@@ -48,7 +19,7 @@ async function run(nodeName, networkInfo, args) {
         api.tx.balances.transfer({ Id: TO.address }, AMOUNT),
         FROM
     );
-  
+
     const newBalance = await api.query.system.account(TO.address);
     console.log('newBalance', newBalance.toString());
